refactor(test): app/fixture 중복 클래스 제거 후 test-support로 통합
TestContainerSupport·ExceptionAssertions·TypeReferenceFrameHandler가
test-support 모듈에 이미 존재하므로 app/fixture 사본을 삭제하고
모든 임포트를 coffeeshout.support.* 로 교체한다.
