+++ backend/app/src/test/java/coffeeshout/patchnote/application/PatchNoteAdminServiceTest.java
@@ -1,6 +1,6 @@
 package coffeeshout.patchnote.application;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
