+++ backend/app/src/test/java/coffeeshout/report/application/ReportFacadeTest.java
@@ -1,6 +1,6 @@
 package coffeeshout.report.application;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import coffeeshout.global.ServiceTest;
