+++ backend/app/src/test/java/coffeeshout/user/infra/persistence/UserEntityTest.java
@@ -1,6 +1,6 @@
 package coffeeshout.user.infra.persistence;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.mockito.Mockito.mock;
 
 import coffeeshout.global.exception.GlobalErrorCode;
