+++ backend/app/src/test/java/coffeeshout/websocket/docs/WsCatalogBuilderTest.java
@@ -1,6 +1,6 @@
 package coffeeshout.websocket.docs;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
